% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACAT.R
\name{ACAT_V}
\alias{ACAT_V}
\title{A set-based test that uses ACAT to combine the variant-level p-values.}
\usage{
ACAT_V(G, obj, weights.beta = c(1, 25), weights = NULL,
  mac.thresh = 10)
}
\arguments{
\item{G}{a numeric matrix or dgCMatrix with each row as a different individual and each column as a separate gene/snp. Each genotype should be coded as 0, 1, 2.}

\item{obj}{an output object of the NULL_Model function.}

\item{weights.beta}{a numeric vector of parameters for the beta weights for the weighted kernels. If you want to use your own weights, please use the “weights” parameter. It will be ignored if “weights” parameter is not null.}

\item{weights}{a numeric vector of weights for the SNPs. When it is NULL, the beta weight with the “weights.beta” parameter is used.}

\item{mac.thresh}{a threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this thrshold.}
}
\value{
p-value of ACAT-V.
}
\description{
A set-based test that uses ACAT to combine the variant-level p-values.
}
\examples{
 data(Geno)
 G<-Geno[,1:100] # Geno is a dgCMatrix of genotypes
 Y<-rnorm(nrow(G)); X<-matrix(rnorm(nrow(G)*4),ncol=4)
 obj<-NULL_Model(Y,X)
 ACAT_V(G,obj)
}
\author{
Yaowu Liu
}
