% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACAT.R
\name{NULL_Model}
\alias{NULL_Model}
\title{Get parameters and residuals from the NULL model}
\usage{
NULL_Model(Y, X = NULL)
}
\arguments{
\item{Y}{a numeric vector of outcome phenotypes.}

\item{X}{a numeric matrix of covariates. X must be full-rank.}
}
\value{
This function returns an object that has model parameters and residuals of the NULL model of no association between genetic variables and outcome phenotypes. After obtaining it, please use ACAT-V function to conduct the association test.
}
\description{
Compute model parameters and residuals for ACAT-V
}
\examples{
 Y<-rnorm(10000)
 X<-matrix(rnorm(10000*4),ncol=4)
 obj<-NULL_Model(Y,X)
}
\author{
Yaowu Liu
}
